xc_domain_save: cache pages mapping
We map the dirty pages, and copy them to secondary machine, and then unmap
it. xc_map_foreign_bulk() is too slow, so we can't use full bandwidth to
transfer dirty pages. In out test, the transfer speed is less than 300Mb/s.
For virtual machine (VM) replication, the transfer speed is very important,
so we should cache pages mapping to map the same page only one time. In
our test, the transfer speed is about 2Gb/s with this patch.

Signed-off-by: Ye Wei <wei.ye1987@gmail.com>
Signed-off-by: Jiang Yunhong <yunhong.jiang@intel.com>
Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>
Signed-off-by: Wen Congyang <wency@cn.fujitsu.com>