Fix a bug in vnif that will cause kernel panic
xenvif_disconnect_suspend() is runned in a work queue, so it can be runned
with xenvif_start_xmit() at the same time. vif->netbk is set to NULL in
xenvif_disconnect_suspend(), and it will be used in xenvif_start_xmit().
Do more check in xen_netbk_queue_tx_skb() to avoid accessing NULL pointer.